Feat: initproc-rs, try fix exit_code

+++ crates/libd/src/syscall/syscall.rs
@@ -52,11 +52,11 @@ pub fn sys_fork() -> isize {
     syscall(SYS_CLONE, [0, 0, 0])
 }
 
-pub fn sys_exec(path: *const u8, argv: *const u8, envp: *const u8) -> isize {
+pub fn sys_exec(path: *const i8, argv: *const i8, envp: *const i8) -> isize {
     syscall(SYS_EXECVE, [path as usize, argv as usize, envp as usize])
 }
 
-pub fn sys_waitpid(pid: usize, exit_code_ptr: *mut usize) -> isize {
+pub fn sys_waitpid(pid: usize, exit_code_ptr: *mut i32) -> isize {
     syscall(SYS_WAIT4, [pid, exit_code_ptr as usize, 0])
 }
